This is a patch I've been using for several months now: it allows xastir to use geotiff files that are already unprojected lat/lon maps.  It simply checks the "PCS" key in the geotif file, and if it's not there then the file is assumed to be in lat/lon and can be used as is without transformation.
It has no effect on maps that are not already in lat/lon, and so will be
invisible to anyone who hasn't needed the ability to read in unprojected
geotiff files.

README.MAPS updated to explain that gdalwarp can be used to convert maps into
lat/lon unprojected maps.  Also added a place to find useful geotiff maps
that need such conversion (FAA sectionals from aviationtoolbox.org).

Tweaked the stuff on shapefile reprojection.  What was there were verbose
snippets from an email exchange between Curt and me.  Made them a bit shorter
and more to-the-point.

The importing of lat/lon geotiff files is significantly slower than the
importing of USGS quads, because the optimized reads that skip over scanlines
based on zoom level don't work.  I intended to figure out how to make them
work, but never did the work.  If folks start using this a lot and get me
to bump it up in priority I can revisit it.