Return "dist-bzip2" to the AUTOMAKE_OPTIONS variable.
it is true that --dist-bzip2 is not a valid automake option, but the
AUTOMAKE_OPTIONS variable doesn't generate --options for automake, it
has an effect on the generated makefile itself.  In this case, dist-bzip2
adds a dist-bzip2 target to the makefile.

As far as I can tell, the folks who are having trouble with this option are
not in fact using automake 1.6.3 or later.  In ha2vr's case it turned out
he was actually using automake 1.4, even though automake 1.9 was supposedly
installed on his system.