Round of RTree usage optimizations.
1) If a shapefile's bounding box is entirely contained in the current viewport,
   don't bother doing an RTreeSearch for visible shapes --- just read the
   file sequentially.  Don't even bother accessing the hash table for the file.

2) Store a time stamp every time a shapefiles hash table entry is accessed.

3) Every hour, walk through the hash table and purge records that have not
   been accessed in 1 hour.  (this needs to become a configurable time)

This should improve memory usage for a few special use cases, and free up
RAM wasted on shapefiles that are not really going to benefit from having
the RTree fast lookup.

I've not run this for long times to see how the memory usage behaves ---
once the RTree is allocated and freed, the processes resident size doesn't
seem to drop down, but it does make the space that was freed available for
indexing a different shapefile.

There are probably many ways to make this even more flexible and automatically
adaptive to the users' usage patterns.  This should be a good start, though.