map_shp.c:  move block of code that accesses rtrees and shapefile hashes to after the block that includes the "return if just indexing".  This lets us skip the hashtable lookup and rtree generation for maps that are merely being index.
shp_hash.c: add a test before RTreeInsertRect to make sure that the bounding
box being passed is valid (i.e. that it won't fail the assertion inside the
RTreeInsertRect routine), and skip the shape if not.

This should fix the crashes Curt was seeing a few days ago when including
some defective Finland shapefiles in his map directory.