Comment out warning when a geotiff file has no PCS.
Geotiffs without PCS (Projected Coordinate System) tags are assumed to be
lat/lon rasters.  Now that this capability is being more widely used,
especially by users outside the US, this warning is just an annoyance.  It was
only there because it was once cause to reject a raster if it had no PCS.