Undo bonehead commit by KM5VY (me).
I had mistakenly thought that "draw_filled" was a new map index parameter, and
tried to get the index_retrieve_from_file function to set defaults for
shapefiles differently.

This was an error, as draw_filled has always existed.

There is no way to "fix up" the index when running with new code -- the user
is simply going to have to go through the map properties and turn Auto on if
he/she wants filled shapefiles to respect the dbfawk file's settings.

The reason I thought the commit was right was that I was testing it on an
index that had been passed through awk to remove the draw_filled variable.
It then inserted reasonable defaults, but that's not the point.

It's unfortunate that a user with an existing batch of filled shapefiles
and dbfawk enabled will wind up surprised that they stopped being filled,
but that's the way it is.