While trying to understand why my TIGER polygon maps have so many identically-named small polygons, I discovered an error in the Xastir_tigerpoly.py script that was propagated from the original GDAL tigerpoly.py script.
The TIGER/Line format has a "Type I" record that links complete chains
(Type 1) to "GT-Polygons" through the "POLYIDL" and "POLYIDR"
attributes.  But according to the TIGER/Line documentation, a Type I
record that contains identical POLYIDL and POLYIDR attributes is a
line that is internal to a polygon, and should not be included in its
boundary.  The original tigerpoly.py script from which
Xastir_tigerpoly.py was derived simply added all lines with either
POLYIDL or POLYIDR to the boundary of polygon POLYID.  That means that
these internal points were not only incorrectly added to the boundary
of POLYID, they were added twice.

A simple test to throw away such "PolyChainLink" data (in the OGR
parlance, "Type I record" in the TIGER parlance) removes this
incorrect assembly of polygon boundaries.

This does *not* fix the issue of there being lots of tiny,
identically-named polygons in a given area (that's just how the data
is), but it DOES correct the "Error in computing fill/hole ring"
warnings I've been getting since forever, every time these
TIGER/Line-converted polygon shapefiles are loaded.  It also fixes the
problem I had trying to import these polygon shapefiles into GRASS,
which noticed and griped angrily about the topological incorrectness
of the data.

Unfortunately, taking advantage of this fix would require regenerating
all the TIGER polygon shapefiles that are up on xastir.tamu.edu ---
probably not worth the effort.  But should there be a new round of
corrected TIGER data that gets processed, the script is now producing
(more) topologically correct shapefiles.

It might be a good idea to feed this back to Frank Warmerdam, on whose
tigerpoly.py script Xastir_tigerpoly.py was based.