Add "-d" option to this script to allow dissolution of common boundaries between polygons with the same Landmark designation.
This is done by scanning all PolyChainLink records, checking that both left
and right polygon have associated AreaLandmark records, and that both
AreaLandmark records point to the same Land record.  If they do, then the
two polygons are dissolved into one.

The option is not the default.

Using this option makes the output shapefile be topologically inequivalent to
the original TIGER/Line data, and as such they are suitable only for display.
Distributing data produced in this manner should be done with a prominient
disclaimer pointing out that the shapefiles are for display purposes, not
serious GIS use.

Dissolving polygons in this way makes the TIGER data look much nicer (but still
pretty ugly) in xastir -- large landmarks no longer have many copies of the
labels, and no longer have ugly internal dashed boundaries that mean nothing.