Implement auto detection of USGS DRG topo maps (the ones actually from USGS, with proper GeoTIFF tags).
Changed map driver interface to take a struct of map flags instead of a single
"draw_filled" integer.

Changed map properties to support setting of Yes/No/Auto for a given map's
"USGS DRG" field.  If Yes, the color map is assumed to be USGS's and the
color settings in Map->Configure USGS DRG are used.  If No, all colors in
the colormap are displayed.  If Auto, the tiff's Image Description tag is
queried and if it begins with "USGS DRG GeoTIFF" then it's as if the setting
was Yes, and if it doesn't (or the tag does not exist) then it is as if the
setting was No.

Added rowcol to map properties dialog to handle the extra buttons without
widening the window.

Cleaned up the code in main.c that locates the filename in a map properties
line.  It was being done by hardcoding the offset in 8 places to 31.  Moved
the magic number to a preprocessor symbol so it only needs to be changed once
now.  There has got to be a cleaner way still, but that's enough for now.

Changed all map drivers to support the new map_data_flags struct instead of
the int draw_filled.

Changed language files.  Not sure I got all of them right.