Move -lgdal to end of LIBS line if it's found.
It is possible for gdal to be using its own internal geotiff library, and
this can interfere with use of an external one in xastir.  Having it later
in the line than geotiff can fix that.

This could cause problems for systems with busted linkers that need gdal to
be earlier than everything it depends on.  If that happens, it might be
necessary to add the contents of "gdal-config --dep-libs" to LIBS after
-lgdal.