Possible fix for bug that prevents dbfawk files with a single field in the dbffields list from working unless a colon is added to the end of the list.
This appears to be a very platform dependent bug, as I have never been able
to reproduce the problem.  But the fix is still valid --- the issue is that
no null terminator is added to a field name unless a colon is found.

This fix doesn't break anything, but I'm unable to verify that it actually
fixes the problem reported, since I have never actually seen that problem
in action.