Fix broken dbffields variables in all dbfawk files.
The dbfawk file is supposed to have semicolon-terminated statements.
Unfortunately, the parser is busted and doesn't actually handle the
case where the semicolons are omitted correctly.  So, for example,

 BEGIN {
  dbfinfo="Blah:Diblah:Diblahdiblah";
  dbffields="Blah:Diblah:Diblahdiblah"
}

winds up getting the newline-space at the end of the line added to
dbffields, and thus in the list of names to be retrieved from the dbf file.
This never matches the last field when applied to the dbf files, which have no
such extra characters in the field names.

This bug is masked by *how* the dbffields variables were actually set in
the included dbfawk files:
 BEGIN {
  dbfinfo="Blah:Diblah:Diblahdiblah";
  dbffields="Blah:Diblah:Diblahdiblah"}

That is, since there was no newline between " and }, dbfawk didn't do the
wrong thing.

Unfortunately, many people have patterned their own dbfawk files after the
ones in the config directory, and so this missing-semicolon deal got propagated
(I am among those propagating it, as all my shape_web dbfawk files omitted
the semicolon, but also put the brace right after the ").  The problem
only gets exposed when the user inserts a newline after the quote.

This commit adds the semicolon (and a newline) after the dbffields closing
quote.

The real fix is to fix the dbfawk/awk parser so it doesn't let syntax errors
like this slip through.