Add comments to show where it is necessary to fix data_out_ax25 so that it isn't confused by leading white-space in commands.
No code changes.  The idea was to document what I know I need to do to fix
the problem that Mike Fenske saw when the \r was added before MYCALL commands.
The quick fix was not to add the \r if sending to an AX25 port, but the clean
fix would be to make data_out_ax25 not care about leading white space.