Add EXTRA_DIST and DIST_SUBDIRS lines so that the shapelib stuff is properly bundled when doing a "make dist" for release.
The Makefile.am in the contrib directory does not actually build any of the
contrib codes.  That's another step for another moment.  I just wanted to
make sure that any make dist that is done is sure to bundle the entire
shapelib directory in order to comply with the terms of the license.

The "Makefile_orig" script in the shapelib/contrib directory could be used
temporarily to build the shapelib tools.  Ultimately, they could be built
by the new makefile, but it's not really the goal of this exercise to do all
that.  In fact, at some point it might be reasonable for xastir's shapelib
directory to contain only the files needed for building the library, but that
will probably require modifying it to be under GPL instead of LGPL.  That's
just a matter of editing some license files, not any code.

Also, removed the shprewind file from the shapelib directory.  This is
a linux binary that for some reason was in the shapelib source tarball I had.
It should never have been added to xastir's CVS repository.