Found yet another annoying issue with libgeotiff build on ubuntu (and probably anywhere where gcc 4 is used).
The problem is that the official tarball has the modification times of
configure.in newer than configure, and a makefile that
regenerates configure whenever it's older than configure.in --- and then runs
it with no arguments.  Since we need to run configure with some
arguments, that completely messes everything up.

Add a "touch configure" command to the commented-out part of the script
that makes configure newer than configure.in and bypasses that part of
the  Makefile.  A kludge upon a kludge upon a kludge.

I missed this subtlety because when I finally tracked down the "ld
-shared" issue, it was in a dirty directory -- one that had already had
configure regenerated and which therefore no longer had the broken
timestamps.