Make DBFAWK trump the gps_flag.
This means that if a file is in the GPS directory and there is
a DBFAWK file to go with it, the DBFAWK file controls the
rendering.  It is still the case that if there is NO dbfawk
file, or if DBFAWK is not enabled, the hard-coded GPS stuff
(i.e. fixed width and line style, color taken from file name)
is used.

Also, change the dbfawk file created upon GPS download to
match the style that the hard-coded GPS stuff would use.