Undo the CoverSplit->l_CoverSplit change made last week.
The real issue was that some repositories have a GDAL that includes GRASS
libraries including a libgrass_rtree that duplicates the rtree functionality.
Xastir's rtree library is precisely the one that GRASS uses (based on
public domain (i.e. uncopyrighted) software from www.superliminal.com), and
this leads to a link conflict that produces an obscure and cryptic error
message at link time.

Renaming the CoverSplit variable was just a band-aid that was suggested for
a trial run while diagnosing the problem, not a recommended fix.