Fix improper probing of curl library when LSB is not being used.
AC_CHECK_LIB can have two or three arguments.  With two arguments it tests
for the presence of a function named in the second argument in the library
named by the first argument, and if found defines HAVE_LIBxxxxx and adds
-lxxxxx to LIBS.

In the three-argument case, it does the probe the same way, but instead of
the default "define HAVE_LIBxxxxx and add -lxxxxx to LIBS" it does what it's
told in the third argument.  That means if you want to rely on HAVE_LIBxxxxx
and LIBS="${LIBS} -lxxxxx" then you have to do it yourself in the third
argument.

The AC_CHECK_LIB for curl was only doing the extra thing, which was to set
"use_curl" to "yes".

I recently discovered this by trying to build and use xastir on a
freshly installed system that had libcurl but no wget (most of my
other systems have both).  The bug in the use of AC_CHECK_LIB meant that
libcurl was found, but never linked in nor was HAVE_LIBCURL ever set in
config.h.

With this fix, curl is now both properly detected *and* the detection properly
communicated to the code.