Added a utility function to return a date/time in w3c timestamp format, required for valid timestamps in kml files.
Rearranged the log station to kml code, and added timestamped placemarks
for each location of a station in a track.  This produces kml output
that validates, and should behave like the "Shroeder" whale shark example
file (moving a station location along its track as a time slider control
is moved in Google Earth).  This has not yet been tested to see if the
behavior is as desired.