Improvements to the tollerance of the db_gis code to problems with connections to databases and improved reporting of problems in making connections.
Added code to add tracks for stations retrieved from the database.
Still needs work, as the existing db.c code for adding stations and
points to the tracks of stations makes different assumptions about
the order in which positions are recieved than the order of records
queried in db_gis.c.