Fix an annoyance related to DF bearing reports.
Sometimes a bad parse of garbled Mic-E data causes Xastir to think that a
D700 (or other mobile) is sending a DF report.  Even if only a single such
garbled transmission is received, the station is shown with DF bearing
information for as long as it's in the database.  Watching a mobile drive
around with randomly moving DF beam lines is distracting.  Not being able to
make it go away without restarting Xastir has been annoying.

Another way for this to happen is for some bonehead (such as KM5VY) to create
a series of DF objects from an external program with his callsign as base name
and sequential numbers to identify the reports (say "KM5VY-1", "KM5VY-2",
etc.) until he happens to hit the SSID of the station that's sending the
reports (say, KM5VY-8).  What happens is the "object" data gets attached to
the Xastir instance itself.  This bogus beam heading shows up forever until
shutting down Xastir.  This happened today on a SAR ELT DF practice.  The
name of the bonehead has not been changed to protect the bonehead.

This commit adds the ability to clear such bogosity.  Now, the "Station Info"
dialog for any station with DF bearing information has a "Clear DF Bearing"
button that will zero out the bearing and NRQ strings for the station,
removing the unwanted (and presumably incorrect) bearing lines until the
station transmits them again (which probably means they're wanted and not
incorrect).