Move "create_garmin_waypoint" call outside of "if (sound_play_prox_message)" conditional.
A local user was completely confused why his garmin wasn't receiving waypoint
data from Xastir, and when I hunted through the coude found that this feature
only works if you have audio alerts for proximity warnings turned on.  This
seems wrong, so I moved the conditionals around so that waypoints are created
if you're within proximity warning distance, whether or not you've actually
got the old audio alarm turned on for proximity warnings.