Add a new nwsc_ddmmyy file with different signature to match the very latest shapefiles.
I am adding this as a new dbfawk instead of modifying the existing one so that
those who don't wish to update their shapefiles will not have to, and those
who have only the new ones will actually be able to use them.  Xastir will
pick the dbfawk file with the matching signature.