Fix for an insidious bug in open2300db2APRS that relied on an unsafe internal behavior of the mysql API.
The function Get_Latest_WX makes 2 SQL queries to the server, one to get the
highest value of the "timestamp" field present in the DB, an another to get
the row with that timestamp.  Only after the first query was the function
calling "row = mysql_fetch_row(&result)".  That function returns a char **
that can be used to pick off the field values.  Somehow, despite never
being called after the second query, the row array still managed (on some
systems) to have the right data in it --- presumably because the API call
was returning a pointer to an internal block of memory that contained the
field values, and on the systems where this worked that block wasn't moving.

When I upgraded my BSD system and recompiled mysql, this started failing, and
open2300db2APRS would segfault.

Calling mysql_fetch_row after the second query fixes it.

While I was at it, I fixed a mistake in some debug output that prints the
number of fields retrieved by the query.