Fix for the problem where Xastir will display black images when compiled against GraphicsMagick with its new default QuantumDepth of 8, or ImageMagick with its optional QuantumDepth of 8.
The problem was that when the image was not a colormapped image (e.g. JPEG
or other full color type), Xastir was assuming (incorrectly) that a Quantum
was the same type as the red, green, and blue components of an XColor, and
using the Quantums directly out of the pixel_pack in a call that required
the latter.  The result was that when pack_pixel_bits returned, the result
was always that the packed pixel bits were always 0x000000, because it was
taking the high-order 8 bits from the unsigned short to which the unsigned
char was being cast.

What was needed was a little block of code that massaged the
pixel_pack values into unsigned shorts at that point if QuantumDepth is 8.

This fix applies the same fix to map_tiger, map_geo and map_WMS, all of which
had duplicated code (ugh) that made the same mistaken assumption of
Quantum/unsigned short equivalence.

This fix does NOT address the problem that exists on a few systems, where
ImageMagick has been compiled with HDRI (High Dynamic Range Image) support.
On those systems, a Quantum is not even an integer type, it's a float or
double.  Dealing with *that* mess will be harder.

Ideally, the duplicated code between map_WMS, map_geo, and map_tiger should
be consolidated according to good software engineering principles.