EXPERIMENTAL COMMIT FOR TESTING PURPOSES.
This commit modifies the weather station code so taht stations which provide
1-hour, 24-hour, and since-midnight rain rates (e.g. Davis through Meteo and
db2APRS, LaCrosse, and Davis APRS Data Logger) do not have those station-
provided data overwritten by Xastir when it tries to recalculate these
quantities from "total rain" (i.e. total rainfall since the weather station
was reset).

The Davis Meteo/LaCrosse code does not even provide "total rainfall" information
(though the database they access does in fact have this information),
and the APRS Data Logger does not provide total rainfall.  All three
provide per-hour and 24-hour rain data, and the Data Logger provides rain-since
-midnight.

What this code does is add a flag to the WeatherRow structure that is set to
1 by stations that provide total rain, and to zero by stations that provide
the rate data instead.  In "cycle_weather", the three rate strings are
only overwritten if the flag is 1, and are left as the station set them
if the flag is 0.

This should address issues of transmitted rain rates being completely
incorrect for Davis and LaCrosse stations.

Needs a lot more testing to assure that there is no impact on other stations.
This code should result in no change in Xastir's behavior for any station
other than Davis or LaCrosse.