Fix two instances of a logical AND being used where a bit-wise AND was intended.
In this case, the test was supposed to be checking that a station posit was
not a third-party packet, but was improperly doing flag && ST_3RD_PT instead
of flag & ST_3RD_PT.

This impacts ONLY speech-enabled band opening alerts.  It is unlikely that
anyone was even using these.  I found this error only because clang spits
out warnings about it.  The specific warning is that one is attempting
to do a logical AND with a constant value (which is more likely an error
than intentional).