Remove a horrid work-around of some bad behavior of Xastir with Kenwood D7* radios with attached GPS.
Back in 2006, several Xastir users noticed that when Xastir was
attached to a Kenwood D700 (and other Kenwood APRS radios) using the
"TNC with GPS on AUX port" interface type, every time Xastir tried to
transmit the TNC would reply with an "Eh?" error when sent the MYCALL
line that precedes a posit.  We realized that there must be some kind of
junk in the TNC buffer that was causing the MYCALL line to not be recognized.
In some cases, this had bad consequences, especially if no previous MYCALL
had succeeded --- the result would be that the TNC would transmit using the
default MYCALL, "NOCALL".

Since we couldn't for the life of us figure out what the junk in the buffer was,
we kludged:  Before sending MYCALL, Xastir would send a carriage return, letting
the "eh?" error happen where it had no consequence.

This was a horrid kludge, as it just swept the real problem under the rug.

The real issue is that "Serial TNC with GPS on AUX port" was written
to support devices like the Kantronics KPC-3+, which requires that
Xastir send a Control-E to the TNC in order to poll it for GPS data.
This is how the KPC-3+ multiplexes the output serial port --- normally
it sends packet data to the serial port, but immediately after
receiving a Control-E from the connected computer, it sends one NMEA
string from the GPS instead.

The Kenwwood radios stream the GPS data and packet data together all the time,
and not only do not require a Control-E to switch between them, they don't
even read the Control-Es.

The "Eh?" errors from the Kenwood TNCs were it complaining that MYCALL was
preceded by all those Control-E characters that Xastir had been sending since
the last transmit.

This commit adds a configuration option to the Serial TNC with GPS on AUX port
set-up dialog so that one can turn this control-E feature on and off as
appropriate.

THIS HAS USER-VISIBLE IMPACT FOR ALL KENWOOD USERS:

If you are currently using Xastir with a Kenwood radio, after doing
this update you MUST open your interface property dialog and de-select
the 'Send Control-E to get GPS data?" check box.  If you do NOT take
this step, you will begin seeing the old "Eh?" problem on your TNC,
and may well transmit your posit as NOCALL.

This may seem an annoyance, but really, this is the fix that we should have put
in place six years ago.