Definitive fix for KAM TNCs.
The KAM apparently does not go into converse mode if one sends "K\r" and
immediately sends the stuff you want to go into the converse.  It requires
a tiny delay between the command to go into converse mode and the data.

I have inserted a 50ms delay after sending the converse command, which should
be insignificant enough that nobody will need it to be a configurable option.

This fix has been confirmed to end the problems with KAM TNCs, and the 50ms
delay was the smallest delay tested.

Also commented out two commands in tnc-startup.kam that do not apparently
exist in the KAM.