Patch from Tom Hayward (KD7LXL):
This patch fixes an issue in the OSM tile download loop. If a tile is
404, 403, or any other error that doesn't return a valid image,
imagemagick (or gm) puts the error info in &exception. Subsequent
iterations read &exception, note the error, and fail to composite any
further tiles. This patch adds GetExceptionInfo() to the end of the
loop, so subsequent iterations do not start off with error info in the
exception struct.