Ifdef out the hack added to main.c to deal with the broken Xorg-server 1.7.5. This server version had a bug that interfered with all motif programs, causing them to grab the cursor and never release it when a context (pop-up) menu was created.  The bug was promptly fixed, but not before several distros locked in version 1.7.5 as their supported version.
The workaround ungrabs the cursor, but has the annoying side-effect of making
the user hold down the button that brings up the menu, rather than allowing
a quick click to access it.

This commit puts the hack into an ifdef, and enables the user to select the
workaround by adding "--with-xorg_175_workaround" to the configure line.
The default behavior is now to NOT use the workaround.

The workaround and configure.ac code to enable it should be removed completely
in one release cycle.  Xorg-server 1.7.5 is long gone, and keeping this code
around is a nod to the fact that some people do not upgrade for years.  One
more release cycle should be enough to forget this mistake ever happened.