Big commit for a little, insidious threading problem.
xa_config.c provides a function called get_user_base_dir(char*) that
returns a pointer to a string that is the concatenation of the base
configuration directory for Xastir (usually ~/.xastir, but selectable by
the user with a command line option or environment variable) and the
argument.

This function was highly thread-unsafe --- it kept a local static char
array that it used to build the concatenated string, then returned a pointer
to the local array.

This function was used in multiple threads, meaning that one thread could
start scribbling into the static array while another was writing into it.
I first tried to fix this by adding a mutex that locked the writes for all
but one thread at a time, but that wasn't even good enough, because after
unlocking the write mutex another thread could write into the buffer while
the first thread was reading the results just produced.

The result of this mess was a highly unpredictable tendency for Xastir to
start producing corrupted paths from this function.  This would show up most
often on multi-core machines with threading libraries that let threads run
simultaneously on different cores, but it could show up on single-core
machines with just the right scheduler.  It is revealed most easily by enabling
multiple logging options (e.g. TNC logging, NET logging, WX logging) and
PNG snapshots all at once.  In testing, I found this sufficient to produce
corrupted file paths in an hour or so.  I'm somewhat surprised that nobody has
ever reported this issue, because I've been seeing it for ages.

This commit replaces the original get_user_base_dir implementation with one
that requires the caller to provide its own buffer, and completely removes
the use of the static buffer.  More than 200 calls to get_user_base_dir
had to be modified, but some simplifcation was also possible --- I was able
to replace a few usages like this:

   xastir_snprintf(dest_string, sizeof(dest_string), "%s", get_user_base_dir("tmp"));

with:
   get_user_base_dir("tmp", dest_string, sizeof(dest_string));

get_user_base_dir still returns the pointer to the destination string, so it
can still be used as an argument to log_data or xastir_snprintf where necessary.

I also replaced several wasteful uses of the function.  There were places where
get_user_base_dir was being called repeatedly with the same argument many
times in a single function.  I replaced this usage in many cases with a single
call to store the answer in a local buffer, and use that repeatedly.

Since implementing this change I have seen Xastir run without producing a single
corrupted path, even with TNC, NET, and WX logging and PNG snapshots turned on.