Remove some unnecessary macros from the probes for Berkeley DB.
The code in acinclude.m4 to probe for Berkeley DB was copied from another
project, and included a macro that tried to guess how to add a "runpath"
to the build.  On most systems this is either -Rlibrarydirectory or
-Wl,rpath,librarydirectory, but the macro in our acinclude.m4 was in fact
correctly identifying the way to do it ONLY on Linux, and then discarding
the information anyway.  Thus, this was a no-op on Linux, and was wrong
elsewhere (but also a no-op).  The technique used to do the probe also broke
when the compiler was clang, but again, to no effect because the result was
discarded anyway.

SO since the result wasn't used, why probe it?  I have removed the macros and
replaced the use of "XASTIR_ADD_LIBRARY_TO" with a simple shell variable
assignment.

While I was at it, I removed the deprecated two-argument call to
AM_INIT_AUTOMAKE.  It has started causing newer versions of automake to
complain, and has been deprecated for some time.