Add a PNG icon for Xastir, converted from the icon.xbm file in the src directory.
I did this because I generally create a desktop shortcut to Xastir on my
laptop, and as of the most recent version of Gnome I am unable to use an
xbm to do so, but a PNG works.

This new icon is installed into ${prefix}/share/xastir/symbols and may be
used by the user to define a custom icon for a desktop shortcut.

Remember, the Xastir project does not directly support distro-specific
(or desktop-specific) packages, but we do want to make sure that those
who generate such packages have the tools and files they need to do so.