Remove the part of this README that shows a warning about a missing ui-mainwindow.h.  That was a relic from when xastir-qt.pro referenced this file.  ui-mainwindow.h is generated by qmake as a result of the mainwindow.ui file being in FORMS, so it is unnecessary to include that file in HEADERS.
Add a note to the end, too.  The "Linux" section seems to reference an attempt
to build xastir-qt on a 32-bit machine when a previous qmake had been done
on a 64-bit machine in the same directory or something.  If one does builds
in clean build directories (out-of-source builds) and doesn't commit generated
files, this is not going to be a problem; qmake should always precede a make
in a clean build directory.
o