Fix another DESTDIR error in a Makefile.am install hook.
config/Makefile.am has a hook that tries to remove a deprecated nwsfz_*.dbfawk
file from config, but does so without the correct use of DESTDIR.

 http://sourceforge.net/p/xastir/bugs/544/

This is the bug report that exposed the original error, and it quite explicitly
states that the error was in both Makefile.am and config/Makefile.am, but
I overlooked the latter one when fixing the bug (because I was going off
email, not the bug report).

Since a devel snapshot was pushed out to fix my last Makefile fix, it would
probably be worthwhile to push another out with this one, especially since CVS
is down.  It would also be worthwhile to add a comment to the bug report
linked above when that happens, because this DESTDIR thing is mainly important
to people creating distro-specific binary packages, and they won't ever use
CVS versions of the code.