Fix for mysterious "relocation R_X86_64_32 against ... cannot be used when making a shared object" error on Ubuntu 14.04.
This was reported before by a user and no solution found.  Turns out the
error is due to a change in how GraphicsMagkic-config reports flags.  It
now sticks a "-fPIE -pie" into ldflags and cflags where it didn't before.

The problem is that Xastir's configure script was incorrectly putting
GraphicsMagick --cflags output into CXXFLAGS but not CFLAGS.   As a result,
the "-fPIE -pie" were being put into LDFLAGS but not CFLAGS, and the
inconsistency is illegal.

Xastir has no C++, so it is odd that we're setting CXXFLAGS anywhere.  Adding
the cflags to CFLAGS fixes this problem.