Remove more vestiges of broken .geos, add two mapquest online map sources.
Thanks to Lee Bengston, K5DAT, for doing the heavy lifting of nailing down
the URLs and checking the terms of use on these two.

"OSM_tiled_mapquest" is the mapquest rendering of OSM data and looks nice.

"OSM_tiled_mapquest-aerial" is satellite imagery of about the resolution of the
old "terraserver" data (i.e. low, maybe 1-meter-per-pixel?), but in color.