Remove probe for XpGetDocumentData from the Xp library.
Xastir has never used this function itself, but on a handful of systems
years ago (2007), it was necessary to link -lXp explicitly because Motif
depended on it.

Some distros in 2015 have completely removed libxp from their package management
systems, and apparently on those systems Motif no longer needs it.  At any
rate, it is wrong for us to be requiring it and explicitly linking against
it.

If we get reports that some systems still require this explict linking to
link in Motif, then we need a smarter probe for Motif, one that does,
effectively:

  Look for XmTextFindString in libXm.  If found, great, add -lXm to link line.
    if fails, try finding XpGetDocumentData in libXp.  If found, add -lXp.
    Try to find XmTextFindString in libXm again.  If found, great, add -lXm.

This will work, because the Autoconf macro AC_CHECK_LIB tries to create a
small program that uses the given function, and to link it against the given
library.  If that works, then the library contains the function.  When it
does this, it uses the complete list of libraries that have been located so
far.  If libXm depends on libXp, the first attempt will fail, but the second
will pass.

I have not done this in this commit, because I think it is probably unnecessary,
and is a pain to do.  If necessary, I'll do it.

Also, remove the hard-coded path to perl in several .pl scripts.  Perl may be
installed somewhere other than /usr/bin on some systems, and indeed it is on
mine.  Using "/usr/bin/env perl" instead assures that perl is found regardless
of where it is in one's path.  This can be a security problem in some scripts,
but these scripts are required at bootstrap time, and I couldn't even bootstrap
now that my system's package manager doesn't create a symlink for
/usr/bin/perl anymore.