Declare get_text_width in draw_symbols.h
clang was puking when draw_symbols.c defined "long get_text_width(...)"
without having declared it as long first.  It was complaining that the
long definition clashed with the implicit "int" implied by the use before
declaration earlier in the file.

Just needed to add the prototype to the header file.