Change two inline functions to static inline
The functions "no_data_selected" in main.c and "max_i" in maps.c are
declared "inline int".  They are the only two functions defined in this
way in all of Xastir.

They no longer work on Cygwin, where the two functions are reported as
undefined at link time.

Adding static to the declaration appears to be the c99 portable
approach.  Current usage appears to rely on older GCC extension paradigms.

Note that there are other inline functions in Xastir, declared in
different ways.  At some point we might run into a similar issue with
those.