Remove old CVS keywords
CVS keywords such as "$Id$" no longer serve any purpose now that we're
using git, and need to be removed.  This commit removes all of the
"$Id$" keywords, and a few other CVS keywords of no current value.

I also had to change some lines of the scripts/lang*.pl files that were
looking for $Id$ tags in their input and replacing them with comments.
The entire intent of this was to add two lines of comments to the top of
the output, and now it just does that, instead of depending on the
presence of the $Id$ keyword.