Fix wrong lower bound on DEFAULT_AIRCRAFT_CLEAR
The configuration setting DEFAULT_AIRCRAFT_CLEAR controls how long
aircraft data will be displayed as fresh before being marked as ghosted.
A setting of "0" means that Xastir will use the same ghost time that is
used for other stations.

Unfortunately, when this feature was added, the xa_config line that
loads the parameter from the xastir.cnf file was written to require a
setting between 1 and 604800, with a default value of 0.  That meant
that if you started Xastir up with a config file that had no setting for
DEFAULT_AIRCRAFT_CLEAR, it would use 0.

From that point forward, every time you start up Xastir it will complain
that the value of DEFAULT_AIRCRAFT_CLEAR is out of bounds, and that it
will instead be using the default value of 0 (which, of course, is out
of bounds):

  xastir.cnf: DEFAULT_AIRCRAFT_CLEAR out-of-range: 0, changing to default: 0

I'm tired of seeing that warning.

This commit changes the bounds so that the actual default value is
considered in bounds for error checking purposes.  It silences the
warning, and changes nothing else.