Added testdbfawk after downloads to test signature
Now using testdbfawk against the downloaded files to test for a valid dbfawk
signature. This works except for the cases where NWS has added a trailing
alpha character to the zip file, but the contents of the zip file do not
contain the trailing alpha character. In any case, should give us a head-up
when the DBF signature changes on the map files.