Remove bashisms from Bourne shell script
The final section of this script was a recent creation meant to check
the DBF signatures of newly-downloaded NWS shapefiles against dbfawk
files in the Xastir config directory.  The technique used to tease apart
filenames (removing a trailing letter) used bash-specific string
substitutions that don't exist in other shells that some operating
systems call "/bin/sh."  Notable examples are Debian (which has /bin/sh
being "dash" instead of "bash"), and FreeBSD (for which /bin/sh is the
original Bourne shell).

This commit removes the bashism, and instead does the string
substitution using sed.