Add git SHA to version displayed in About box
Xastir has always shown the current version of the code in the dialog
displayed by Help->About.  This version is usually X.Y.Z, with Z an
odd number for "development" versions that have not been released, and
even for official releases.

Thing is, during development phases, "2.0.9" isn't really
representative of what version of the code you're running, because the
code may be different from day to day.

This commit adds the git commit hash (SHA-1) to the version string
displayed here.  When building, we check the source tree for the
presence of a ".git" directory.  If it exists, we're building out of a
git repository and not a tarball, so we run a command to get the
current commit hash.  If we don't find ".git" in the source tree, we
conclude the user is building from a release tarball, and skip it.

So now, if you're building from a git repo, you'll see the version
reported as VX.Y.Z (<sha>).

I find this very useful, because it lets me know whether or not I'm
actually running a binary built from the current state of my git
repo.  It's especially important when working on branches or trying to
resolve pull requests.

Note that this commit adds a script "XastirGitStamp.sh" that actually
gets the git SHA for src/Makefile.am.  It is NOT installed by make
install, and so is not added to scripts/Makefile.am in the
dist_scripts directory.  This script simply has to live in the source
tree, and is only ever referenced from there at build time.

Additionally, it has always bothered me that "compiledate.c"
doesn't always get rebuilt when you make.  It only gets rebuilt if a
dependency changes --- so it isn't really the compile date/time, it's
some date/time after the last code change.

For development versions, sometimes the code doesn't change, but
scripts or Makefiles do, and this forces a rebuild.  But the rules for
compiledate.c *don't* force a recreation of the file in these
instances.

This commit therefore also makes compiledate.c a ".PHONY" target,
which will therefore always be regenerated by make, no matter what.
This assures that developers will *always* have the correct git sha in
their Help->About box that was actually used to build the code.