Remove references to "mkinstalldirs"
In commit bfcea98, Curt added code to bootstrap to copy
"mkinstalldirs" from a system automake directory because problems had
surfaced with automake 1.9 on his system.

Thing is, mkinstalldirs stopped being installed by automake 1.8
precisely because it was no longer the preferred way to work.  I am
guessing that if reference to mkinstalldirs had been removed from
Makefile.am at the time, so that we stopped demanding that it exist,
this would not have been needed.

At any rate, bootstrap only looks in two specific directories for
mkinstalldirs, and some package managers on Mac don't put things in
those directories --- so users are getting confusing messages about
bootstrap not working.

This commit removes the copy of mkinstalldirs from bootstrap, and also
removes mention of mkinstalldirs from all the Makefile.ams.  It's been
almost 10 years since automake deprecated mkinstalldirs, so it should
be safe to stop referring to it.

You will want to rm the copy of mkinstalldirs that is in your source
tree after this commmit, so that automake stops looking at it.  I have
gone ahead and removed mkinstalldirs from .gitignore, so it'll show up
as an untracked/uncommitted file in your git checkout now.  Please
don't accidently commit it and push!  Just remove it.