Add a "wxnowsrv.pl" script
Users have been asking for Xastir to be able to read Wxnow.txt
formatted files from weather station control software for many years.
See, for example,
https://sourceforge.net/p/xastir/feature-requests/304/, a feature
request from 2004.

Since Xastir has a "Network WX" interface type, and that interface
type is supposed to be a server that feeds data that is almost the
same as the contents of Wxnow.txt, I chose to write a quick-and-dirty
perl daemon that munges the contents of Wxnow.txt and feeds them to
Xastir in exactly the same way that the Davis "db2APRS" and LaCrosse
"open2300db2APRS" programs do.

To use this script, run it in the background with two or three command
line arguments.

   /path/to/wxnowsrv.pl wxnowpath polltime [port]

where wxnowpath is the full path to the Wxnow.txt file, polltime is
the polling interval in seconds, and port is the optional port number
to use for the daemon (default is 5500 if this is not given).

Wxnow.txt will then be checked every polltime seconds, and if it has
changed, all connected clients will be given a string of weather data
in the same format that db2APRS and open2300db2APRS produce.

Once the script is running, add a Networked WX interface to
Xastir (Interface->Interface Control->Add) and give the IP address of
the machine on which the script is being run and the port number
chosen.  Xastir will then show your current weather data in View->Own
Weather Data.

If you change your station type to one of the weather station
choices ("Station position w/weather" or "Station position, zulu
date-time and weather" in File->Configure->Defaults), your weather
information will be transmitted to RF and APRS-IS as well.