Remove cvs2cl script
This script was used by Xastir developers to generate the ChangeLog
file from CVS history.  It is irrelevant now, as we a) haven't updated
ChangeLog since 2012, because it's an old fashioned way to record
changes, and b) don't use CVS anymore.

Incidentally, the last time we updated ChangeLog was around the time
we added cvs2cl to the scripts directory.