Make some perl scripts less system-specific
Most perl scripts in the scripts directory have "#!" lines that hard
code the location of perl as /usr/bin/perl.  This is not the universal
location of perl for all systems that Xastir runs on, although it
works on Linux.  Perl may be installed elsewhere on other systems,
such as on BSD, OS X with a package manager, etc.

Replacing "/usr/bin/perl" with "/usr/bin/env perl" fixes this, making
the script find perl based on the user's PATH variable instead of hard
coding it.