Remove outdated README.win32
README.win32 was the original documentation of how to run Xastir on
Windows under Cygwin.  It was replaced by README.CYGWIN in 2010,
and this later file was recently given an extensive update.

README.win32 was horribly outdated and referenced ancient distros like
"lintronix" that no longer exist and just confuse the issue --- it
should have been removed when README.CYGWIN was created.  I am
removing it so that the Cygwin instructions appear in only one place.