Fix build breakage, remove references to removed file
In commit 1d6e920 I removed "README.win32" because it was an ancient
relic that had ceased to be maintained -- and therefore ceased to be
correct.

I forgot to remove all the places that the build system makes
reference to it, and so Curt's Jenkins CI testing failed.

This should fix it all.  While I was fixing that, I cleaned out other
references to README.win32 that were scattered around.