Unbreak build after IPv6 UDP merge
The code I just merged to resolve pull request #16 introduced a new compilation
failure on some versions of Linux.  The code used AI_DEFAULT, which is
documented in the getipnodebyname man page as being valid, but which is
apparently NOT defined in a number of linux platform's netdb.h.

The linux systems that don't have AI_DEFAULT defined include ubuntu 14.04
(which I run on my laptop), and at least one version of SUSE (on which Curt
runs a Jenkins build to test that Xastir continues to compile after pushes).