Shut up warnings
Clang is rather promiscuous with warnings, and Xastir was throwing a
number of them.

This commit fixes a few things like removing unused variables (that are never
used, even inside ifdefs), extraneous parentheses in
conditionals (which clang stupidly warns about), use of "\0" (a
two-character array) where '\0' (a single character) was intended, and
use of "(char) 0x00" where a null pointer was intended.

It also fixes two incorrect uses of strncat as
strncat(dest,src,sizeof(src)), when src is already a null-terminated
string.  The purpose of strncat is to specify a maximum number of
characters to copy to prevent buffer overruns, and using the size of
the source array is pointless, and makes it no safer than using
"strcat(dest,src)".  Replaced with
strncat(dest,src,sizeof(dest)-strlen(dest)).

Also fixes one incorrect usage of strerror_r.  This was an odd one,
because there's an ifdef HAVE_POSIX_STRERROR_R that is never set by
any configure probe, but then the "else" for the ifdef USES
strerror_r, but incorrectly.  So if we *did* probe for a
posix-compatible strerror_r and didn't find it, then this code
wouldn't compile.  And if strerror_r is found, and an error occurs,
the wrong thing would be printed --- it was being used as if it were
strerror, which returns a string pointer, but strerror_r returns an int.