Fix more incorrect strncat uses
strncat(dst,src,len) is supposed to be used in preference to
strcat(dst,src) because the latter can cause a buffer overrun if src
is longer than the remaining space in dst.

Using strncat(dst,src,strlen(src)) or strncat(dst,src,sizeof(src)) is
no better than using strcat --- it says to copy all the chars of src,
even if there's not room in the destination.

Most uses of strncat in xastir are the more correct
strncat(dst,src,sizeof(dest)-strlen(dest)-1) (when dest is an
explicitly declared char array), but the code is full of instances
where the number of chars to transfer is the full length of the source
buffer, not a restricted length to prevent buffer overruns.

This commit fixes only a few of the wrong ones, inspired by my
warning-correcting commit 6a55d02b (clang actually emits a warning for
exactly "strncat(dest,src,sizeof(src))").

Unfortunately, there are a number of uses elsewhere where a single
character of a larger string is pulled out and transfered using
strncat(dest,src,1), and fixing *those* to be
buffer-overrun-safe *and* still be doing the right thing would take
more careful work than I have time for now.

In doing this, I found one place in dbfawk.c where there was
potentially (but not practically) insufficient space allocated to
store an anticipated file name.  A buffer was allocated with 7 more
chars than the source file name, and then ".dbfawk" (7 chars) was
appended to the source file name.  That appears not to leave room for the
null terminating character.  In reality, though, the ".dbf" or ".shp"
extension of the source file name was stripped off first, so this
error never would be encountered.  Still, it's wrong, so I fixed it.