Update CONTRIBUTING document
This document still had detailed instructions for how to create patch
files and submit them to Xastir developers via email.  This is no
longer the recommended way for users to submit suggested code changes.

I have written up a detailed thumbnail sketch of how to do the same
thing via GitHub pull requests, which is a better, more transparent,
and easier way for user contributions to be tracked, reviewed, and
added to Xastir.