Update README, use markdown
README had an *enormous* section on OS-specific hints that was
unbelievably old, with references to some systems' information having
been updated as recently as 2002.  It celebrated the new discovery
that Xastir could build on Solaris 7!  It documented ancient versions
of OSen that nobody uses anymore.

OS-specific information like this doesn't belong in the top-level
README.  It's hard enough to keep up with new os updates on the wiki,
where people are constantly updating things based on their immediate
experiences.

Also, README is displayed on the top-level page of the github site for
Xastir.  As such it should be a little more readable.  So in addition
to ripping out verbose stuff that is irrelevant, I've added markdown
"markup" in the hopes that it will be more readable.