Copy README to README.md
For some reason I don't understand, I was unable to rename README to
README.md and get the build to work, even when changing all references
to README in Makefile.am to README.md --- automake complained that a
required file did not exist, and refused to create Makefile.in.

So I am making a COPY of README into README.md, so that Github will
render the markdown properly.

This is icky, as it will require that we keep README and README.md in
sync.

README is installed by make install, README.md is not.  Github will
display README.md in preference to README on the home page when both exist.