Update help-* files with "What's New"
A long-neglected release task is to add a "What's New in Xastir X.Y"
section to the help file (which is displayed when the user chooses
"Help->Help Index" in Xastir).  This has not been updated since
release 2.0.1 in 2012, and prior to that had not been updated since
release 1.7 in 2005.

I've stuck in a list of 2.0.1->2.0.9 highlights.

I also added a copy of this ENGLISH text into the French, German,
Italian, and Spanish help files, and removed the very, very old
"What's New" sections from those files (some even had "What's New in
Xastir 1.0" sections.  If you are fluent in one of these
languages, translation assistance would be greatly appreciated.  It is
also highly likely that the actual translated text in these documents
is very outdated.

The Dutch and Portuguese help files have never been translated, and so
are just straight copies of the English file.