Clean up .gitignore
When we do the next stable release, we're going to do it by creating a
release branch and *committing normally ignored files to it*.  These
normally ignored files are all things produced by bootstrap, so a user
doesn't have to run bootstrap to build from a stable release tarball.

To do this requires that .gitignore be edited on the release branch so
git stops ignoring those files.

I have cleaned up gitignore so that the files we need to remove on a
release branch are all up at the top and commented so we can spot them
easily.  I also removed a lot of stuff that was unnecessarily
enumerating explicit files in subdirectories that would already be
covered by a global pattern (e.g. Makefile.in and src/Makefile.in are
both handled just by putting "Makefile.in" in the gitignore).

This should make creation of a functional release branch and Github
release tag easier.  I want to do that SOON.