Clean up README section about mailing lists.
The mailman page for the two Xastir mailing lists shows:

  Xastir - APRS client software discussion
  Xastir-dev - APRS client software developer's discussion

That makes it pretty clear that the second is intended
for *developer's* discussion, not general user discussion.

A little repetition in the README led to a bit of confusion today
about whether the README was referring to xastir or xastir-dev at one
point.  I have rearranged the text and added a comment about
xastir-dev that I hope will be clearer.