Fix source path in xastir.spec.in
The xastir.spec.in had a path to Xastir sources that was pretty much
expecting what just got imported from CVS.  New releases of Xastir
have a different path, so I'm updating it.

On the other hand, the presence of .spec files in the Xastir source
tree is a relic of very early days, when some Xastir developers built
RPMs and LSB builds of Xastir themselves.  Now, Xastir packages are
NOT maintained by the Xastir developers, and are the responsibility of
distro-specific package maintainers.  It could be argued that this
spec file doesn't belong here anymore, and neither do some of the
RPM-related relics in the scripts directory (none of which are in
current use).