Update README.developers.md with revised release process
I got some push back about the naming scheme for release tags and the
files that Github would produce for them.

A minor modification of the process (possibly exploiting a poorly
documented feature of the Github archive API) seems to make it
possible to resolve the annoying "Double-Xastir" file naming issue and
so I'm documenting both the change in the release process and the
change in the release tarball URL in README.developers.md.

I'm going to give this a few days to percolate back through the same
package maintainers who already tested out the "fake release" process
I went though for 2.0.8 on Wednesday and were happy with it, to make
sure they are still happy with this modification.  I am confident that
this will satisfy both the folks who were OK with it before and the
folks who had objections.