Fix up outdated URLs in INSTALL
INSTALL still read like it did back in the day when all the libraries
Xastir depended on needed to be built from source.  The URLs for many
such packages were way out of date, as were the version numbers.  I've
simplified this by removing the "current" version numbers and fixing
up some of the URLs.  Most of the packages can be installed via
pacakge management these days, and we say so.