Address compiler warnings in rtree
The rtree library had a number of issues that cause GCC 6 to whine.
For starters, there were two places where an assertion had a condition
like
   assert (pointer >= (void *) 0)

instead of a more normal
   assert (pointer != NULL)

GCC on Raspbian Stretch complained about using an ordered comparision
of pointer with null pointer.

There was a variable that was declared, assigned to, and never used.
This is of course harmless, but the modern trend in C compilers is to
warn about just about everything that *might* be a mistake.

The huge number of warnings coming out of an Xastir build on newer
OSen is becoming a problem, and we should address them.

While I was here, I re-indented some files to remove tabs, and also
removed instances of the "register" storage class specifier.  This is
an ancient construct that is basically ignored by modern compilers.
It is deprecated in C++, and removed in C++17, for example.