Clean up compiler warnings in db.c
These were all "variable declared and set, but never used" warnings.
GCC 6.x whines incessantly about that sort of thing.

In some cases, because it was possible we might someday want the data
that is collected by those variables, I just commented them out.  In
others, since it appeared to be leftover debugging cruft, I just took them
out.

gcc 8 finds further issues, where a pointer-to-char array was used
like this:
  "if (station_struct->char_pointer == '\0')"
to check for an empty string.  This is wrong.  What was really
intended was:
  "if (station_struct->char_pointer[0] == '\0')"

Finally, clang on OS X was warning about this construct:

   if (!strncasecmp(....) == 0)

According to C precedence rules, this is literally equivalent to:
  if (  (!strncasecmp(...))  == 0)

but it could easily have been intended to be the very different:
  if (  !(strncasecmp(...)  == 0)   )

Clang whines that you should put parens in to make absolutely clear
what the intent was.  strncasecmp returns zero if the string matches,
and non-zero (either positive or negative) if there is no match.

It does not make sense to take the logical not of the integer return
value, so I changed it to this:
  if (  strncasecmp(...)  != 0   )

That is, if there is no match, we execute the body of the if.