Silence GCC warnings from draw_symbols.c
There were a lot of warnings here about variables that were set but
never used.  Most appeared to be relics of experimental code that
either got deleted or commented out.  Some of them were variables that
were used to compute something that would be needed for a feature that
never got implemented (e.g. oval PHG rings that are tilted to indicate
directivity).

For most of these, I've removed the variable.  For the "tilt" thing, I
just inserted a debug statement that would only be printed if debug
level is 4095 (all bits on), which is essenntially drinking from the
firehose and something that will never be done.  But it makes GCC shut
up.

There were a couple of X calls whose returned values were save in
variables but never used --- that's because they're manage widgets,
and specified as being managed by some other widget, and the return
value is unneeded.  To silence the annoying GCC warning, the thing to
do is not bother saving the return value.