Fix a really dumb error in nwsozddmmyy_14.dbfawk
Somehow, despite having tested my work and seen it to function
properly, I managed to commit the wrong version of this file.  And had
an error (just fixed in previous commit) in the Makefile so the wrong
version didn't get installed and overwrite the one I had that worked.
I don't know how that happened, but I chalk it up to trying to do too
many things at once, too quickly.