Clean up LICENSE file
I found the LICENSE file a little hard to read, and thought it could
use a little reworking to clarify which parts of Xastir are GPL and
which are not, and to make it really clear that some of the later text
applies only if one is distributing prebuilt binaries.

I have gone through the file and explicitly enumerated all the files
in the Xastir source tree that are NOT GPLv2, and pointed to the files
in question for the exact license terms.

Closes #31