Add Magick-config info to the FAQ
A user had a question today about Xastir building with ImageMagick
that appeared to be because he had Magick-config in /usr/local/bin,
but didn't have /usr/local/bin in his path.

The FAQ entry on "Why isn't ImageMagick recognized..." had a lot of
RPM-specific reasons why this might not happen, but nothing about this
extremely simple error.  Now it has the simple answer first.