Add a background parallel download manager
The DownLoad Manager (DLM) will transfer up to 8 tiles at a time.
It can also be used as a generic file downloader.

OSM tiles are now downloaded to the cache in the background with
up to 8 tiles being transferred at a time. The most recently requested
tiles are downloaded first so that the area in view is updated in a
timely manner. We then go back to catch up on any remaining tiles.

As tiles become available, the map is updated on screen.

Tiles are cached for up to 7 days. If an older tile is in the cache,
a download is queued, and the old tile is used until the updated
version is available.