Fix ancient error in festival SPEAK_ID handling
Ever since it was first introduced in commit 374e8cf34 in 2002, there
was a cut-and-paste error in the code that introduced the ability to
speak an ID through festival.

This error was just reported by Mike Finch <finchm@swva.net> who found
the exact line where the code was wrong.  While the code was correctly
extracting a configuration option from the .cnf file, it was setting
the wrong boolean.