Fix typo in strncasecmp calls
There is a paradigm in db.c of calling strncasecmp in conditionals as:
  if (strncasecmp(stringarray,"text",4)==0)
to if stringarray is the the 4-character string "text".

There were two cases in a single conditional where there was a slight oops:
  if (strncasecmp(stringarray,"text",4==0))
that is, where "4==0" was the third argument of strncasecmp instead of
4 being the arg, and the return value compared to zero.

This got flagged as a warning when I was building just now, and it is
clearly an error.  I fixed it.