Remove gdal and ogr support
Xastir has never really used GDAL for raster support, and had only
preliminary, somewhat kludgy OGR support for a small subset of OGR
vector types.

As a consequence of having partial support for these types, Xastir
added a fairly heavy-weight dependency on the GDAL library, and there
is a function in libgdal that has the same name as one in libgeotiff.
As a result, a kludge was introduced (in 2005, in commit 0e8675bf3)
that duplicated the GTIFProj4FromLatLong function of libgeotiff, but
renamed to "my_GTIFProj4FromLatLong".

The problem with this approach was that it introduced a direct use of
the proj.4 API into Xastir, and the "proj" project has deprecated the
API used by that call.

By removing GDAL support, we can also remove the my_GTIF... function
and go back to relying on the function provided by libgeotiff ---
which causes Xastir to have only INDIRECT dependence on proj.  Thus,
we push off all questions of proj deprecation onto the libgeotiff
project, which has already handled the issue in their upcoming 1.5.x
release.