Remove shapelib contrib programs dependent on proj
The shpcntrd, shpdata, shpproj, and shpwkb programs are all fairly
useless and depend on the old proj.4 API that has been deprecated in
proj 6.0 and will be removed in proj 7.0.

There is no value to having these ancient contrib programs in our
internal shapelib, because they just introduce a use of proj that we
have to maintain.  In fact, the shapelib project has removed them.

All of the functions that had been previously provided (badly) by
these programs can be had using ogr2ogr and other tools provided by
gdal.

The only reason we have an internal shapelib is because once upon a
time it could not reasonably be assumed that a package manager had a
version of shapelib in it, and we wanted shapefile support to be a
default in Xastir.  The decision was taken to include shapelib source
code in Xastir, and to use it only if no externally installed shapelib
was found.  This was a bad choice and I would like to undo it now that
shapelib is in just about every package management system that exists,
and we can simply make it a required TPL.  But that's an issue for
another day.  In the meantime, there is no value to maintaining tools
that even the shapelib project has abandoned.

Resolves issue #47.

?