Make Xastir build reproducible
Insertion of the compilation date and time into the Xastir binary
breaks binary reproducibility, a feature desired (maybe even demanded)
by some Linux package management systems.

This commit removes the compile date and time from the "compiledate.c"
and instead only inserts the git sha-1 (which it was already doing)
if, and only if, we're building out of a git repo.

Doing this makes the binary produced from any Xastir code base
identical (i.e. with the same MD5 checksum) to any other binary built
on the same system with the same libraries from the same code base.

This closes issue #51.