Remove all vestiges of old tigermap code
The server that used to serve the census Tiger maps went down many
years ago, but all the code that enabled displaying it and configuring
the map features remained --- all of this code was unusable.

There is now a WMS server available that does the same thing, and it's
in the code base as a .geo file.  It does NOT use any of the
configuration that was enabled by the code base put in place for the
old server.

I am removing all vestiges of this old code to resolve issue #12.