Remove OpenTrac decoding routines
OpenTrac was a hot new thing back around 2003.  It is mostly dead now.

And worse, the OpenTrac code is now generating a lot of warnings from
GCC 8, some of which look like they might actually point to potential
errors.  Since the code may be borken, and nobody uses the OpenTrac
protocol anymore, this commit simply removes the functions that handle
decoding those packets in the interests of making the cleanup of
warnings (issue #24) easier.

Resolves #56.