Remove commented-out code
I am currently working with some users of the DRAWS hat to figure out
why Xastir is unable to read GPS data from GPSD when this GPS device
is in use, even though other programs seem to be able to read the data
just fine.

In the course of investigating, one user found some commented-out code
in interface.c and thought it prudent to try to uncomment it, because
the comments implied that Xastir wouldn't work with GPSD after version
2.90 without that code being uncommented.

In fact, the code was moved long ago in commit 00eef061, but rather
than removing the code from the original location, it was just
commented out.

There is no good reason to leave commented-out code in a file when we
have a perfectly good version control system to allow us to back out
changes when needed.  The presence of such code can be confusing to
people reading the code, and it is generally the case that when such
code stays in for many years, it wouldn't work anyway if it were
uncommented.