Add support for GNGGA and GNRMC sentences
GLONASS receivers send $GNGGA sentences instead of $GPGGA sentences, and
$GNRMC instead of $GPRMC.  But the only difference between these
sentences is the GN instead of GP --- the data is identical.

This appears to be the reason that users of the DRAWS hat are unable
to use this device with Xastir, even though it appears to work with
all other programs that consume NMEA sentences.

This commit attempts to recognize GGA and RMC sentences irrespective
of whether they begin with "$GP" or "$GN".  It is tested so far only
with GPS receivers, and it doesn't break those.  I'm adding this on a
branch so that users of the DRAWS hat can test it out before we
actually merge into master.