Removing the internal Shapelib tree
Removing the stale version of Shapelib that we compiled in if Shapelib
was not found on the system. Most systems have Shapelib available now
or can obtain it easily, so keeping an old copy of Shapelib in the
Xastir sources was deemed unworkable. This should also fix the
"distclean" issue that Tom Russo explained on the Xastir list on
04/15/2019: I encountered that bug personally while trying to build
inside a "build" directory.