Remove use of "values" and "values.pl"
The script snippets "values" and "values.pl" were generated from
"values.in" and "values.pl.in", respectively, at configure time.  A
handful of scripts then included (via "source" for shell or "require"
for perl) these snippets for no reason other than to obtain the
install prefix given in the "--prefix" option to configure.

The result was that the scripts themselves would ONLY work if run
from the installation directory, OR if one ill-advisedly builds Xastir
in its own source tree --- if one followed the recommended procedure
of building out-of-source, the scripts would be broken because
values (or values.pl) don't live in the same directory as the script.

This commit does away with values and values.pl entirely, and has
configure scribble the prefix directly into the scripts themselves ---
in exactly the same manner as it had been scribbling into the values*
files.

I am also removing xastir-migrate.sh and xastir-fixcfg.sh.  These two
scripts were added in July of 2003 when the Xastir project reorganized
how Xastir installs itself.  At the time anyone who was running older
code would have had to move a lot of data and fix up their .cnf files
to upgrade to the new code and not lose settings and maps.  These
scripts did that for them so they didn't have to figure anything out.
But nobody is running 2003-vintage Xastir anymore, and if they were,
the right answer to "why isn't Xastir finding my maps and other
configuration" would be "Because you need to wipe out your old install
and start over" not "run this 16 year old script."