Remove hard-coded perl and bash paths
A lot of scripts in this directory had "#!/usr/bin/perl" hard coded at
their tops.  This is wrong, because perl may not be installed
there (it isn't on a lot of non-Linux systems).  I have replaced the
hard-coded path with "/usr/bin/env perl" which will locate the perl
binary in the user's PATH.

Some scripts in this directory already had that.  Those are scripts
that I actually use myself on a system that has perl in
/usr/local/bin, and so I had already fixed them.

Another alternative to using /usr/bin/env would be to create a
configure test to find the perl binary, then substitute it in a .in
version of the script.  That is more work, for negligible benefit.